backend/requirements.txt
absl-py==2.1.0
addict==2.4.0
basicsr==1.4.2
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
facexlib==0.3.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
gfpgan==1.3.8
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.5
kiwisolver==1.4.7
lazy-loader==0.4
llvmlite==0.41.1
lmdb==1.6.2
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.2
-e git+https://github.com/xinntao/Real-ESRGAN.git@a4abfb2979a7bbff3f69f58f58ae324608821e27#egg=realesrgan
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
six==1.17.0
sympy==1.13.3
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.2
tifffile==2023.7.10
tomli==2.2.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.6
yapf==0.43.0
zipp==3.20.2
